src/requirements.txt
yarl<1.2
https://github.com/Rapptz/discord.py/archive/rewrite.zip#egg=discord.py[voice]
asyncio
requests
cogs
gitpython
liburl
selenium
wolframalpha
wikipedia
googletrans
whois
